+++ boot/espressif/hal/include/bootloader_wdt.h
@@ -6,3 +6,4 @@
 #pragma once
 
 void bootloader_wdt_feed(void);
+void bootloader_config_wdt(void);
