+++ boot/espressif/hal/src/esp32s2/bootloader_init.c
@@ -26,6 +26,7 @@
 #include "soc/extmem_reg.h"
 #include "soc/io_mux_reg.h"
 
+#include "bootloader_wdt.h"
 #include "hal/wdt_hal.h"
 
 #include "esp32s2/rom/cache.h"
@@ -33,12 +34,7 @@
 #include "esp32s2/rom/spi_flash.h"
 #include "esp32s2/rom/uart.h"
 
-esp_image_header_t WORD_ALIGNED_ATTR bootloader_image_hdr;
-
-void bootloader_clear_bss_section(void)
-{
-    memset(&_bss_start, 0, (&_bss_end - &_bss_start) * sizeof(_bss_start));
-}
+extern esp_image_header_t WORD_ALIGNED_ATTR bootloader_image_hdr;
 
 static void bootloader_reset_mmu(void)
 {
@@ -49,14 +45,6 @@ static void bootloader_reset_mmu(void)
     /* normal ROM boot exits with DROM0 cache unmasked,
     but serial bootloader exits with it masked. */
     REG_CLR_BIT(EXTMEM_PRO_ICACHE_CTRL1_REG, EXTMEM_PRO_ICACHE_MASK_DROM0);
-}
-
-esp_err_t bootloader_read_bootloader_header(void)
-{
-    if (bootloader_flash_read(ESP_BOOTLOADER_OFFSET, &bootloader_image_hdr, sizeof(esp_image_header_t), true) != ESP_OK) {
-        return ESP_FAIL;
-    }
-    return ESP_OK;
 }
 
 static void update_flash_config(const esp_image_header_t *bootloader_hdr)
@@ -130,28 +118,6 @@ static esp_err_t bootloader_init_spi_flash(void)
 
     update_flash_config(&bootloader_image_hdr);
     return ESP_OK;
-}
-
-void bootloader_config_wdt(void)
-{
-    wdt_hal_context_t rtc_wdt_ctx = {.inst = WDT_RWDT, .rwdt_dev = &RTCCNTL};
-    wdt_hal_write_protect_disable(&rtc_wdt_ctx);
-    wdt_hal_set_flashboot_en(&rtc_wdt_ctx, false);
-    wdt_hal_write_protect_enable(&rtc_wdt_ctx);
-
-#ifdef CONFIG_ESP_MCUBOOT_WDT_ENABLE
-    wdt_hal_init(&rtc_wdt_ctx, WDT_RWDT, 0, false);
-    uint32_t stage_timeout_ticks = (uint32_t)((uint64_t)CONFIG_BOOTLOADER_WDT_TIME_MS * rtc_clk_slow_freq_get_hz() / 1000);
-    wdt_hal_write_protect_disable(&rtc_wdt_ctx);
-    wdt_hal_config_stage(&rtc_wdt_ctx, WDT_STAGE0, stage_timeout_ticks, WDT_STAGE_ACTION_RESET_RTC);
-    wdt_hal_enable(&rtc_wdt_ctx);
-    wdt_hal_write_protect_enable(&rtc_wdt_ctx);
-#endif
-
-    wdt_hal_context_t wdt_ctx = {.inst = WDT_MWDT0, .mwdt_dev = &TIMERG0};
-    wdt_hal_write_protect_disable(&wdt_ctx);
-    wdt_hal_set_flashboot_en(&wdt_ctx, false);
-    wdt_hal_write_protect_enable(&wdt_ctx);
 }
 
 static void bootloader_init_uart_console(void)
